Improve testing
